added basic d6 dice rolling and reorganized files
