+++ options/distinguishing-features.js
@@ -1,0 +1,45 @@
+const distinguishingFeatures = {
+  Scar: [
+    'Pockmarked cheeks',
+    'Gash across left eye',
+    'Burned hand',
+    'Noose mark around throat',
+    'Missing ear',
+    'Frequestly broken nose',
+  ],
+  Tattoo: [
+    'Ocean waves along arms',
+    'Holy symbol on rump',
+    'Intricate maze on back',
+    'Strange arcane markings on chest',
+    'Broken chains on wrists',
+    'Kill count on forearm',
+  ],
+  Piercing: ['Ear', 'Septum', 'Neck', 'Eyebrow', 'Tongue', 'Belly-button'],
+  Scent: [
+    'Freshly cut grass',
+    'Stale meat',
+    'Ocean mist',
+    'Rotten eggs',
+    'Oil and grease',
+    'Smoke and ash',
+  ],
+  Accent: [
+    'Slow and deliberate',
+    'Quick and precise',
+    'Thinks out loud',
+    'Lazy Drawl',
+    'Rough whisper',
+    'Bellowing and commanding',
+  ],
+  Affectation: [
+    'Plays with hair/beard',
+    'Flips a coin to make decisions',
+    'Wears two hats',
+    'Imitates a different Species',
+    'Always snacking',
+    'Whistles or sings when nervous',
+  ],
+};
+
+module.exports = distinguishingFeatures;
