+++ options/eye-color.js
@@ -1,0 +1,23 @@
+const eyeColor = {
+  human: ['Black', 'Green', 'Blue', 'Brown', 'Hazel', 'Amber'],
+  aelf: ['White', 'Gold', 'Jade', 'Sapphire', 'Amethyst', 'Ruby', 'Black'],
+  duardin: [
+    'Molten Gold',
+    'Rich Green',
+    'Pale Blue',
+    'Dark Brown',
+    'Bright Orange',
+    'Flame Red',
+    'Coal Black',
+  ],
+  sylvaneth: [
+    'Snow White',
+    'Ice Blue',
+    'Bright Yellow',
+    'Verdant Green',
+    'Soft Orange',
+    'Deep Brown',
+    'Night Black',
+  ],
+};
+module.exports = eyeColor;
