+++ options/eye-type.js
@@ -1,0 +1,40 @@
+const eyeType = [
+  'Mornful',
+  'Bitter',
+  'Haunted',
+  'Sad',
+  'Cold',
+  'Hollow',
+  'Smiling',
+  'Hopeful',
+  'Bright',
+  'Caring',
+  'Trusting',
+  'Warm',
+  'Piercing',
+  'Striking',
+  'Searching',
+  'Hard',
+  'Calculating',
+  'Dead',
+  'Soft',
+  'Deep',
+  'Hagard',
+  'Old',
+  'Young',
+  'Misty',
+  'Wild',
+  'Narrow',
+  'Weary',
+  'Uncaring',
+  'Intense',
+  'Stormy',
+  'Sparkling',
+  'Gentle',
+  'Wise',
+  'Cunning',
+  'Burning',
+  'Mesmerising',
+];
+
+module.exports = eyeType;
