+++ options/hair-color.js
@@ -1,0 +1,39 @@
+const hairColor = {
+  human: [
+    'White Blond',
+    'Golden Blond',
+    'Black',
+    'Brown',
+    'Auburn',
+    'Red',
+    'Grey',
+  ],
+  aelf: [
+    'White',
+    'Silver',
+    'Honey Blond',
+    'Chestnut Brown',
+    'Mahogany Brown',
+    'Deep Red',
+    'Black',
+  ],
+  duardin: [
+    'White',
+    'Grey',
+    'Golden',
+    'Copper',
+    'Bronze',
+    'Dark Brown',
+    'Black',
+  ],
+  sylvaneth: [
+    'Newly Budding',
+    'Pale Yellow',
+    'Soft Green',
+    'Verdant Green',
+    'Golden Brown',
+    'Wilted',
+    'Dead',
+  ],
+};
+module.exports = hairColor;
