+++ src/character_features.js
@@ -1,150 +1,10 @@
 const { d6 } = require('./dice');
 const { randomInt } = require('./util');
 
-const distinguishingFeatures = {
-  Scar: [
-    'Pockmarked cheeks',
-    'Gash across left eye',
-    'Burned hand',
-    'Noose mark around throat',
-    'Missing ear',
-    'Frequestly broken nose',
-  ],
-  Tattoo: [
-    'Ocean waves along arms',
-    'Holy symbol on rump',
-    'Intricate maze on back',
-    'Strange arcane markings on chest',
-    'Broken chains on wrists',
-    'Kill count on forearm',
-  ],
-  Piercing: ['Ear', 'Septum', 'Neck', 'Eyebrow', 'Tongue', 'Belly-button'],
-  Scent: [
-    'Freshly cut grass',
-    'Stale meat',
-    'Ocean mist',
-    'Rotten eggs',
-    'Oil and grease',
-    'Smoke and ash',
-  ],
-  Accent: [
-    'Slow and deliberate',
-    'Quick and precise',
-    'Thinks out loud',
-    'Lazy Drawl',
-    'Rough whisper',
-    'Bellowing and commanding',
-  ],
-  Affectation: [
-    'Plays with hair/beard',
-    'Flips a coin to make decisions',
-    'Wears two hats',
-    'Imitates a different Species',
-    'Always snacking',
-    'Whistles or sings when nervous',
-  ],
-};
-
-const eyeType = [
-  'Mornful',
-  'Bitter',
-  'Haunted',
-  'Sad',
-  'Cold',
-  'Hollow',
-  'Smiling',
-  'Hopeful',
-  'Bright',
-  'Caring',
-  'Trusting',
-  'Warm',
-  'Piercing',
-  'Striking',
-  'Searching',
-  'Hard',
-  'Calculating',
-  'Dead',
-  'Soft',
-  'Deep',
-  'Hagard',
-  'Old',
-  'Young',
-  'Misty',
-  'Wild',
-  'Narrow',
-  'Weary',
-  'Uncaring',
-  'Intense',
-  'Stormy',
-  'Sparkling',
-  'Gentle',
-  'Wise',
-  'Cunning',
-  'Burning',
-  'Mesmerising',
-];
-
-const eyeColor = {
-  human: ['Black', 'Green', 'Blue', 'Brown', 'Hazel', 'Amber'],
-  aelf: ['White', 'Gold', 'Jade', 'Sapphire', 'Amethyst', 'Ruby', 'Black'],
-  duardin: [
-    'Molten Gold',
-    'Rich Green',
-    'Pale Blue',
-    'Dark Brown',
-    'Bright Orange',
-    'Flame Red',
-    'Coal Black',
-  ],
-  sylvaneth: [
-    'Snow White',
-    'Ice Blue',
-    'Bright Yellow',
-    'Verdant Green',
-    'Soft Orange',
-    'Deep Brown',
-    'Night Black',
-  ],
-};
-
-const hairColor = {
-  human: [
-    'White Blond',
-    'Golden Blond',
-    'Black',
-    'Brown',
-    'Auburn',
-    'Red',
-    'Grey',
-  ],
-  aelf: [
-    'White',
-    'Silver',
-    'Honey Blond',
-    'Chestnut Brown',
-    'Mahogany Brown',
-    'Deep Red',
-    'Black',
-  ],
-  duardin: [
-    'White',
-    'Grey',
-    'Golden',
-    'Copper',
-    'Bronze',
-    'Dark Brown',
-    'Black',
-  ],
-  sylvaneth: [
-    'Newly Budding',
-    'Pale Yellow',
-    'Soft Green',
-    'Verdant Green',
-    'Golden Brown',
-    'Wilted',
-    'Dead',
-  ],
-};
+const distinguishingFeatures = require('../options/distinguishing-features');
+const eyeType = require('../options/eye-type');
+const eyeColor = require('../options/eye-color');
+const hairColor = require('../options/hair-color');
 
 function addDiceResults(dice) {
   return [...Array(dice)]
